UI: add frontend routes and refactor into cleaner components
