+++ package.json
@@ -49,6 +49,8 @@
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
     "react-dropzone": "^4.2.9",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "request": "^2.85.0",
     "request-promise-native": "^1.0.5",
     "sprintf-js": "^1.1.1"
