+++ ui/app/index.jsx
@@ -1,14 +1,55 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
+
 import App from './App.jsx';
+import Home from './pages/Home.jsx';
+import Detail from './pages/Detail.jsx';
+import Error404 from './pages/Error404.jsx';
+
 import registerServiceWorker from './registerServiceWorker';
 
-import APIClient from '../../client'
+import APIClient from '../../client';
 
 const api = new APIClient()
 
-ReactDOM.render(
-  <App api={api}/>,
-  document.getElementById('root')
-);
+//get config as starter
+api.getConfig( config => {
+
+  // config and global props to pass down to children
+  const {
+    site,
+    social,
+    isTestnet,
+    defaultNetwork,
+    version
+  } = config
+
+  const generalProps = {
+    api,
+    site,
+    social,
+    isTestnet,
+    defaultNetwork,
+    version
+  }
+
+  ReactDOM.render(
+      <Router>
+        <App {...generalProps}>
+          <Switch>
+            <Route path="/" component={ props =>
+              <Home {...generalProps} {...props}/>
+            }/>
+            <Route path='/detail/:hash' component={Detail} />
+            <Route component={Error404} />
+          </Switch>
+        </App>
+      </Router>
+    ,
+    document.getElementById('root')
+  )
+})
+
+
 registerServiceWorker();
