+++ ui/app/pages/Detail.jsx
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+const Detail = ({
+  hash
+}) => (
+  <div>
+    {hash}
+  </div>
+)
+
+export default Detail
